refactor: 구간 관련 클래스 패키지 위치 이동
