+++ src/main/java/subway/line/SectionRepository.java
@@ -1,0 +1,7 @@
+package subway.line;
+
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface SectionRepository extends JpaRepository<Section, Long> {
+
+}
