--- src/main/java/subway/section/SectionRepository.java
@@ -1,12 +1,0 @@
-package subway.section;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.List;
-
-public interface SectionRepository extends JpaRepository<Section, Long> {
-
-    List<Section> findAllByLineIdOrderById(Long lineId);
-
-    void deleteAllByLineId(Long lineId);
-}
