+++ src/test/java/subway/line/LineAcceptanceTest.java
@@ -9,7 +9,6 @@ import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.http.HttpStatus;
-import subway.section.SectionRepository;
 import subway.station.StationRepository;
 
 import java.util.HashMap;
